+++ generators/app/templates/api/blueprints/findone.js
@@ -1,6 +1,11 @@
 var _ = require('lodash');
 var actionUtil = require('sails/lib/hooks/blueprints/actionUtil');
 
+var takeAliases = _.partial(_.pluck, _, 'alias');
+var populateAliases = function (model, alias) {
+  return model.populate(alias);
+};
+
 /**
  * Find One Record
  * GET /:model/:id
@@ -10,14 +15,15 @@ var actionUtil = require('sails/lib/hooks/blueprints/actionUtil');
 module.exports = function (req, res) {
   _.set(req.options, 'criteria.blacklist', ['limit', 'skip', 'sort', 'populate', 'fields']);
 
-  var fields = req.param('fields') ? req.param('fields').split(',') : false;
+  var fields = req.param('fields') ? req.param('fields').replace(/ /g, '').split(',') : [];
+  var populate = req.param('populate') ? req.param('populate').replace(/ /g, '').split(',') : [];
   var Model = actionUtil.parseModel(req);
   var PK = actionUtil.requirePk(req);
-  var findQuery = actionUtil.populateEach(Model.findOne(PK), req);
+  var findQuery = _.reduce(_.intersection(populate, takeAliases(Model.associations)), populateAliases, Model.findOne(PK));
 
   findQuery
     .then(function (_record) {
-      var record = fields ? _.pick(_record, fields) : _record;
+      var record = fields.length > 0 ? _.pick(_record, fields) : _record;
       return record ? res.ok(record) : res.notFound();
     })
     .catch(res.serverError);